Move Android types to separate files
