+++ src/BlazorWebView/src/Maui/Android/BlazorWebViewHandler.Android.cs
@@ -73,16 +73,14 @@ namespace Microsoft.AspNetCore.Components.WebView.Maui
 				throw new InvalidOperationException($"Can't start {nameof(BlazorWebView)} without native web view instance.");
 			}
 
-			var assetConfig = Services!.GetRequiredService<BlazorAssetsAssemblyConfiguration>()!;
-
 			// We assume the host page is always in the root of the content directory, because it's
 			// unclear there's any other use case. We can add more options later if so.
 			var contentRootDir = Path.GetDirectoryName(HostPage!) ?? string.Empty;
 			var hostPageRelativePath = Path.GetRelativePath(contentRootDir, HostPage!);
 
-			var compositeFileProvider = new MauiAssetFileProvider(contentRootDir);
+			var mauiAssetFileProvider = new MauiAssetFileProvider(contentRootDir);
 
-			_webviewManager = new AndroidWebKitWebViewManager(this, NativeView, Services!, MauiDispatcher.Instance, compositeFileProvider, hostPageRelativePath);
+			_webviewManager = new AndroidWebKitWebViewManager(this, NativeView, Services!, MauiDispatcher.Instance, mauiAssetFileProvider, hostPageRelativePath);
 			if (RootComponents != null)
 			{
 				foreach (var rootComponent in RootComponents)
@@ -100,79 +98,5 @@ namespace Microsoft.AspNetCore.Components.WebView.Maui
 
 		protected virtual WebChromeClient GetWebChromeClient() =>
 			new WebChromeClient();
-
-		public partial class MauiAssetFileProvider : IFileProvider
-		{
-			private string _contentRootDir;
-
-			public MauiAssetFileProvider(string contentRootDir)
-			{
-				_contentRootDir = contentRootDir;
-			}
-
-			public IDirectoryContents GetDirectoryContents(string subpath)
-				=> new AndroidMauiAssetDirectoryContents(Path.Combine(_contentRootDir, subpath));
-
-			public IFileInfo GetFileInfo(string subpath)
-			{
-				var path = Path.Combine(_contentRootDir, subpath);
-				try
-				{
-					var file = Android.App.Application.Context.Assets!.Open(path);
-					Func<Stream> stream = () => Android.App.Application.Context.Assets!.Open(path);
-					return new AndroidMauiAssetFileInfo(Path.GetFileName(path), stream);
-				}
-				catch (Exception)
-				{
-					return new NotFoundFileInfo(Path.GetFileName(subpath));
-				}
-			}
-
-			public IChangeToken? Watch(string filter)
-				=> null;
-		}
-
-		class AndroidMauiAssetFileInfo : IFileInfo
-		{
-			private Func<Stream> _factory;
-
-			public AndroidMauiAssetFileInfo(string name, Func<Stream> factory)
-			{
-				Name = name;
-				_factory = factory;
-				using var stream = factory();
-				using var memoryStream = new MemoryStream();
-				stream.CopyTo(memoryStream);
-				memoryStream.Seek(0, SeekOrigin.Begin);
-				Length = memoryStream.Length;
-			}
-
-			public bool Exists => true;
-			public long Length { get; }
-			public string PhysicalPath { get; } = null!;
-			public string Name { get; }
-			public DateTimeOffset LastModified { get; } = DateTimeOffset.FromUnixTimeSeconds(0);
-			public bool IsDirectory { get; } = false;
-
-			public Stream CreateReadStream()
-				=> _factory();
-		}
-
-		class AndroidMauiAssetDirectoryContents : IDirectoryContents
-		{
-			public AndroidMauiAssetDirectoryContents(string subpath)
-			{
-			}
-
-			List<AndroidMauiAssetFileInfo> files = new List<AndroidMauiAssetFileInfo>();
-
-			public bool Exists => false;
-
-			public IEnumerator<IFileInfo> GetEnumerator()
-				=> throw new NotImplementedException();
-
-			IEnumerator IEnumerable.GetEnumerator()
-				=> throw new NotImplementedException();
-		}
 	}
 }
